climai-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
# MongoDB driver (with SRV support)
pymongo[srv]==4.9.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to version compatible with passlib; newer versions may lack __about__ metadata

# Data stack versions compatible with Python 3.13 on Windows (prebuilt wheels)
# Loosen pins to avoid building from source
pandas>=2.2.3,<3
scikit-learn>=1.5,<2
joblib>=1.3
cloudpickle>=3.0
google-generativeai>=0.8.0
